testutils cleanup
