+++ testutils/transactions.go
@@ -1,15 +1,4 @@
-/*
- * Test helpers.
- */
 package testutils
-
-import (
-	"encoding/json"
-	"testing"
-
-	"github.com/flashbots/rpc-endpoint/types"
-	"github.com/flashbots/rpc-endpoint/utils"
-)
 
 // Test tx for bundle-failed-too-many-times MM1 fix
 var TestTx_BundleFailedTooManyTimes_RawTx = "0x02f9019d011e843b9aca008477359400830247fa94def1c0ded9bec7f1a1670819833240f027b25eff88016345785d8a0000b90128d9627aa40000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000016345785d8a000000000000000000000000000000000000000000000000001394b63b2cbaea253a00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000002000000000000000000000000eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee0000000000000000000000006b175474e89094c44da98b954eedeac495271d0f869584cd00000000000000000000000086003b044f70dac0abc80ac8957305b6370893ed0000000000000000000000000000000000000000000000d3443651ba615f6cd6c001a011a9f58ebe30aa679783b31793f897fdb603dd2ea086845723a22dae85ab2864a0090cf1fcce0f6e85da54f4eccf32a485d71a7d39bc0b43a53a9e64901c656230"
@@ -21,31 +10,6 @@ var TestTx_MM2_RawTx = "0xf86980850e5b35485d8252089409f427f1bd2d7537a02812275d03
 var TestTx_MM2_From = "0x7AaBc7915DF92a85E199DbB4B1D21E637e1a90A2"
 var TestTx_MM2_Hash = "0xc543e2ad05cffdee95b984df20edd2e38e124c54461faa1276adc36e826588c9"
 
-var RpcEndpointUrl string // set by test init()
-
-func SendRpcAndParseResponse(req *types.JsonRpcRequest) (*types.JsonRpcResponse, error) {
-	return utils.SendRpcAndParseResponseTo(RpcEndpointUrl, req)
-}
-
-func SendRpcAndParseResponseOrFailNow(t *testing.T, req *types.JsonRpcRequest) *types.JsonRpcResponse {
-	res, err := SendRpcAndParseResponse(req)
-	if err != nil {
-		t.Fatal("sendRpcAndParseResponse error:", err)
-	}
-	return res
-}
-
-func SendRpcAndParseResponseOrFailNowString(t *testing.T, req *types.JsonRpcRequest) string {
-	var rpcResult string
-	resp := SendRpcAndParseResponseOrFailNow(t, req)
-	json.Unmarshal(resp.Result, &rpcResult)
-	return rpcResult
-}
-
-func SendRpcAndParseResponseOrFailNowAllowRpcError(t *testing.T, req *types.JsonRpcRequest) *types.JsonRpcResponse {
-	res, err := SendRpcAndParseResponse(req)
-	if err != nil {
-		t.Fatal(err)
-	}
-	return res
-}
+// Test tx for MM2 fix
+var TestTx_CancelRelay_RawTxInitial = "0x02f90197010c851a13b86000851a13b860008302134094def1c0ded9bec7f1a1670819833240f027b25eff80b90128d9627aa400000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000008b6f8cc1baffdb0400000000000000000000000000000000000000000000000000084b1e89ebc7c5000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000020000000000000000000000006b175474e89094c44da98b954eedeac495271d0f000000000000000000000000eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee869584cd00000000000000000000000086003b044f70dac0abc80ac8957305b6370893ed0000000000000000000000000000000000000000000000c2b17f6d6d619e0daac080a032854f4cd2ef72624bd4b24ac933677fa82d68915fb40cbc5ded938e55e0d60ba03d668997706bf52d9d6fd06de1e0b3ecb5de60c78faba54ee6f61c017236b530"
+var TestTx_CancelRelay_RawTxCancel = "0x02f86d010c851caf4ad000851caf4ad0008302134094c0e1142e97a9679fa0f9c13067af656cc24753738080c080a04db949f68b275646517c8225f2b783476579bfc58b6e6af9c0ead0b41f3f5fb1a0540bbff674288973d5c82c25bc7b8fe42343c4d64d85f79a9e21ab6698738103"
